attention-ocr/chkpoint/requirements.txt
torch
torchvision
ignite
Pillow
captcha
pycrypto